OIPA/requirements.txt
lxml==3.4.0
Django==1.9.9
django-tastypie==0.12.0
python-dateutil==2.2
python-mimeparse==0.1.4
defusedxml==0.4.1
contexttimer==0.3.0
mechanize==0.2.5
django-suit==0.2.11
objgraph==1.8.1
ujson==1.33
simplejson
redis==2.10.3
rq==0.4.6
rq-scheduler==0.5.0
django-rq==0.7.0
djangorestframework
future==0.15.2
